rounds model spec corrected
