+++ app/models/challenge.rb
@@ -1,4 +1,3 @@
 class Challenge < ActiveRecord::Base
-  has_many :crawl_challenges
-  has_many :crawls, through: :crawl_challenges, dependent: :destroy
+  has_many :rounds
 end
