+++ app/models/pub.rb
@@ -1,4 +1,3 @@
 class Pub < ActiveRecord::Base
-  has_many :crawl_pubs
-  has_many :crawls, through: :crawl_pubs, dependent: :destroy
+  has_many :rounds
 end
